fix: adjust tooltip
